Initial LSTM/BN/LN implementation
