+++ scripts/train.sh
@@ -17,7 +17,7 @@
 echo "Container nvidia build = " $NVIDIA_BUILD_ID
 
 DATA_DIR=${1:-"/datasets/LibriSpeech"}
-MODEL_CONFIG=${2:-"configs/default_rnnt.toml"}
+MODEL_CONFIG=${2:-"configs/rnnt.toml"}
 RESULT_DIR=${3:-"/results"}
 CHECKPOINT=${4:-"none"}
 CREATE_LOGFILE=${5:-"true"}
@@ -27,6 +27,7 @@ PRECISION=${8:-"fp16"}
 EPOCHS=${9:-3000}
 SEED=${10:-6}
 BATCH_SIZE=${11:-8}
+EVAL_BATCH_SIZE=${11:-8}
 LEARNING_RATE=${12:-"0.0003"}
 GRADIENT_ACCUMULATION_STEPS=${13:-1}
 LAUNCH_OPT=${LAUNCH_OPT:-"none"}
@@ -60,6 +61,7 @@ fi
 
 CMD=" train.py"
 CMD+=" --batch_size=$BATCH_SIZE"
+CMD+=" --eval_batch_size=$EVAL_BATCH_SIZE"
 CMD+=" --num_epochs=$EPOCHS"
 CMD+=" --output_dir=$RESULT_DIR"
 CMD+=" --model_toml=$MODEL_CONFIG"
@@ -68,12 +70,12 @@ CMD+=" --seed=$SEED"
 CMD+=" --optimizer=adam"
 CMD+=" --dataset_dir=$DATA_DIR"
 #CMD+=" --val_manifest=$DATA_DIR/librispeech-dev-clean-wav.json"
-CMD+=" --train_manifest=$DATA_DIR/librispeech-train-clean-100-wav.json,$DATA_DIR/librispeech-train-clean-360-wav.json,$DATA_DIR/librispeech-train-other-500-wav.json"
-CMD+=" --val_manifest=$DATA_DIR/librispeech-dev-clean-wav.json"
-#CMD+=" --train_manifest=$DATA_DIR/librispeech-dev-clean-wav-8.json"
+#CMD+=" --train_manifest=$DATA_DIR/librispeech-train-clean-100-wav.json,$DATA_DIR/librispeech-train-clean-360-wav.json,$DATA_DIR/librispeech-train-other-500-wav.json"
+CMD+=" --val_manifest=$DATA_DIR/librispeech-dev-clean-wav-8.json"
+CMD+=" --train_manifest=$DATA_DIR/librispeech-dev-clean-wav-8.json"
 CMD+=" --weight_decay=0"
-CMD+=" --save_freq=10000"
-CMD+=" --eval_freq=3000000"
+CMD+=" --save_freq=1000"
+CMD+=" --eval_freq=10000"
 CMD+=" --train_freq=25"
 #CMD+=" --lr_decay"
 CMD+=" --gradient_accumulation_steps=$GRADIENT_ACCUMULATION_STEPS "
